+++ R/initfuncs.R
@@ -1,83 +1,86 @@
-init_authcode <- function()
+init_authcode <- function(init_args)
 {
     stopifnot(is.list(self$token_args))
     stopifnot(is.list(self$authorize_args))
 
-    if(!requireNamespace("httpuv", quietly=TRUE))
-        stop("httpuv package must be installed to use authorization_code method", call.=FALSE)
-
-    # browse to authorization endpoint to get code
-    auth_uri <- httr::parse_url(private$aad_endpoint("authorize"))
-
     opts <- utils::modifyList(list(
-        client_id=self$client$client_id,
-        response_type="code",
-        redirect_uri="http://localhost:1410/",
-        resource=self$resource,
-        scope=paste(self$scope, collapse=" "),
-        client_secret=self$client$client_secret,
-        login_hint=self$client$login_hint,
-        state=paste0(sample(letters, 20, TRUE), collapse="") # random nonce
+        resource=if(self$version == 1) self$resource else self$scope,
+        tenant=self$tenant,
+        app=self$client$client_id,
+        password=self$client$client_secret,
+        username=self$client$login_hint,
+        aad_host=self$aad_host,
+        version=self$version
     ), self$authorize_args)
 
-    auth_uri$query <- opts
-    redirect <- httr::parse_url(opts$redirect_uri)
-    host <- if(redirect$hostname == "localhost") "127.0.0.1" else redirect$hostname
-    code <- listen_for_authcode(auth_uri, host, redirect$port)
+    auth_uri <- do.call(build_authorization_uri, opts)
+    redirect <- httr::parse_url(auth_uri)$query$redirect_uri
+
+    code <- init_args$auth_code
+    if(is.null(code))
+    {
+        if(!requireNamespace("httpuv", quietly=TRUE))
+            stop("httpuv package must be installed to use authorization_code method", call.=FALSE)
+
+        code <- listen_for_authcode(auth_uri, redirect)
+    }
 
     # contact token endpoint for token
-    access_uri <- private$aad_endpoint("token")
-    body <- c(self$client, code=code, redirect_uri=opts$redirect_uri, self$token_args)
+    access_uri <- private$aad_uri("token")
+    body <- c(self$client, code=code, redirect_uri=redirect, self$token_args)
 
     httr::POST(access_uri, body=body, encode="form")
 }
 
 
-init_devcode <- function()
+init_devcode <- function(init_args)
 {
-    # contact devicecode endpoint to get code
-    dev_uri <- private$aad_endpoint("devicecode")
-    body <- private$build_access_body(list(client_id=self$client$client_id))
-
-    res <- httr::POST(dev_uri, body=body, encode="form")
-    creds <- process_aad_response(res)
-
-    # tell user to enter the code
-    cat(creds$message, "\n")
+    creds <- init_args$device_creds
+    if(is.null(creds))
+    {
+        creds <- get_device_creds(
+            if(self$version == 1) self$resource else self$scope,
+            tenant=self$tenant,
+            app=self$client$client_id,
+            aad_host=self$aad_host,
+            version=self$version
+        )
+        cat(creds$message, "\n")
+    }
 
     # poll token endpoint for token
-    access_uri <- private$aad_endpoint("token")
+    access_uri <- private$aad_uri("token")
     body <- c(self$client, code=creds$device_code)
 
     poll_for_token(access_uri, body, creds$interval, creds$expires_in)
 }
 
 
-init_clientcred <- function()
+init_clientcred <- function(init_args)
 {
     # contact token endpoint directly with client credentials
-    uri <- private$aad_endpoint("token")
+    uri <- private$aad_uri("token")
     body <- private$build_access_body()
 
     httr::POST(uri, body=body, encode="form")
 }
 
 
-init_resowner <- function()
+init_resowner <- function(init_args)
 {
     # contact token endpoint directly with resource owner username/password
-    uri <- private$aad_endpoint("token")
+    uri <- private$aad_uri("token")
     body <- private$build_access_body()
 
     httr::POST(uri, body=body, encode="form")
 }
 
 
-init_managed <- function()
+init_managed <- function(init_args)
 {
     stopifnot(is.list(self$token_args))
 
-    uri <- private$aad_endpoint("token")
+    uri <- private$aad_uri("token")
     query <- utils::modifyList(self$token_args,
         list(`api-version`=getOption("azure_imds_version"), resource=self$resource))
 
@@ -85,8 +88,12 @@ init_managed <- function()
 }
 
 
-listen_for_authcode <- function(url, localhost="127.0.0.1", localport=1410)
+listen_for_authcode <- function(remote_url, local_url)
 {
+    local_url <- httr::parse_url(local_url)
+    localhost <- if(local_url$hostname == "localhost") "127.0.0.1" else local_url$hostname
+    localport <- local_url$port
+
     # based on httr::oauth_listener
     info <- NULL
     listen <- function(env)
@@ -106,7 +113,7 @@ listen_for_authcode <- function(url, localhost="127.0.0.1", localport=1410)
     on.exit(httpuv::stopServer(server))
 
     message("Waiting for authentication in browser...\nPress Esc/Ctrl + C to abort")
-    httr::BROWSE(url)
+    httr::BROWSE(remote_url)
 
     while(is.null(info))
     {
@@ -154,3 +161,4 @@ poll_for_token <- function(url, body, interval, period)
     message("Authentication complete.")
     res
 }
+