+++ tests/testthat/test02_v2_token.R
@@ -183,3 +183,20 @@ test_that("Certificate authentication works",
     tok <- get_azure_token(res, tenant, cert_app, certificate=cert_file, version=2)
     expect_true(is_azure_token(tok))
 })
+
+
+test_that("Standalone auth works",
+{
+    res <- "https://management.azure.com/.default"
+
+    auth_uri <- build_authorization_uri(res, tenant, native_app, version=2)
+    code <- AzureAuth:::listen_for_authcode(auth_uri, "http://localhost:1410")
+    tok <- get_azure_token(res, tenant, native_app, version=2, auth_code=code, use_cache=FALSE)
+    expect_identical(tok$hash(), aut_hash)
+
+    creds <- get_device_creds(res, tenant, native_app, version=2)
+    cat(creds$message, "\n")
+    tok2 <- get_azure_token(res, tenant, native_app, auth_type="device_code", version=2, device_creds=creds,
+        use_cache=FALSE)
+    expect_identical(tok2$hash(), dev_hash)
+})